#31 part 2
All References Set, UI-Scene should now work in every Level (just the levels for now)
BUT! There is a weird issue with the Code in Line 158 of the SceneManager (commented for now) which freezes Unity whenever it gets processed in Playmode! Further investigations needed here!
When issue fixed: All UI Prefabs of Levels 1-5 can be deleted and line 158 in SceneManager will load the new (now fully functional) UI in